Account handling
SQL
Changed the DB to add user account related stuff and for better readability.

Python
 lots of edit to server.py to handle users and new db field names.
removed an unused function (get_class)
Added register and login routes
register adds an entry in the user table
while login verifies if the credentials match the data in the table and if they do return the user id.

JS
Added class caching to reduce server requests, the class names are stored in the localStorage until the page is either reloaded or a class is added.
fixed a bug that caused jquery to be unable to read the default date when creating an assignment.

added new error alerts to indicate when a field is incorrectly filled.

added register and login functions that do what you would expect.

CSS
improved upon the color picker to make it actually good and nice

HTML
added navbar links for logging in and for logging out, the logout link only appears if the user is logged in.

created login and register pages that contains the login/register "form" (in quotes cause it's never "summited")
the actual posting and filtering and stuff happens in the js functions of the same name.

added a logout page that resets the local storage (to logout) and redirects to the index.